requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.42.0
pydantic==2.10.4
python-dotenv==1.0.1
pydantic-settings==2.7.1
python-multipart
slowapi
PyJWT